+++ src/Core/Content/App/Lifecycle/Event/AppDeletedEvent.php
@@ -2,12 +2,16 @@
 
 namespace Swag\SaasConnect\Core\Content\App\Lifecycle\Event;
 
+use Shopware\Core\Framework\Api\Acl\Permission\AclPermissionCollection;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Event\ShopwareEvent;
+use Swag\SaasConnect\Core\Framework\Webhook\Hookable;
 use Symfony\Contracts\EventDispatcher\Event;
 
-class AppDeletedEvent extends Event implements ShopwareEvent
+class AppDeletedEvent extends Event implements ShopwareEvent, Hookable
 {
+    public const NAME = 'app_deleted';
+
     /**
      * @var string
      */
@@ -24,6 +28,11 @@ class AppDeletedEvent extends Event implements ShopwareEvent
         $this->context = $context;
     }
 
+    public function getName(): string
+    {
+        return self::NAME;
+    }
+
     public function getAppId(): string
     {
         return $this->appId;
@@ -32,5 +41,21 @@ class AppDeletedEvent extends Event implements ShopwareEvent
     public function getContext(): Context
     {
         return $this->context;
+    }
+
+    /**
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.ReturnTypeHint
+     */
+    public function getWebhookPayload(): array
+    {
+        return [];
+    }
+
+    /**
+     * @phpcsSuppress SlevomatCodingStandard.Functions.UnusedParameter
+     */
+    public function isAllowed(string $appId, AclPermissionCollection $permissions): bool
+    {
+        return $appId === $this->getAppId();
     }
 }