+++ src/Core/Content/App/Lifecycle/Event/AppUpdatedEvent.php
@@ -4,4 +4,10 @@ namespace Swag\SaasConnect\Core\Content\App\Lifecycle\Event;
 
 class AppUpdatedEvent extends AppInstalledEvent
 {
+    public const NAME = 'app_updated';
+
+    public function getName(): string
+    {
+        return self::NAME;
+    }
 }